// dts/Enum.ts
enum MTR {
    Log = 'mtr-log',
    Info = 'mtr-info',
    Error = 'mtr-erro',
    Preference = 'mtr-preference',
    Hotkey = 'mtr-hotkey',
    Path = 'mtr-path',
    UpdateArchive = 'mtr-update-archive',
    LoadArchive = 'mtr-load-archive',
    DeleteArchive = 'mtr-delete-archive',
}

enum RTM {
    Log = 'rtm-log',
    Dialog = 'rtm-dialog',
    Hotkey = 'rtm-hotkey',
    EnableHotkey = 'rtm-enable-hotkey',
    LoadArchive = 'rtm-load-archive',
    DeleteArchive = 'rtm-delete-archive',
}

enum Hotkey {
    Save = 'hotkey_save',
    Display = 'hotkey_display',
}


export {
    MTR,
    RTM,
    Hotkey
}